+++ changelog.md
@@ -10,6 +10,10 @@ example:
 + 0.1.1 means amir committed his first commit to version 1.  
 ## Change Log
 
++ 0.4.3.4:
+- comments for teller and score were added.
++ 0.4.3.3:
+- comments and documents were fully added to game of Court Piece
 + 0.4.3.2:
  - comments and documents were fully added to the game of Hearts
 + 0.4.3.1: